requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.3.0
tushare>=1.2.0
akshare>=1.8.0
talib-binary
scipy>=1.7.0 